refactor: move AndroidLogger to infrastructure/logger (#24)
